API

Documentation for GLMCopua.jl's functions.

Index

GLMCopula.GLMCopulaVCModel

GLMCopula.fit!

Functions

GLMCopula.fit! — Function

fit!(gcm::GLMCopulaARModel)

Fit an GLMCopulaVCModel object by MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.ρ, gcm.σ2.

fit!(gcm::GLMCopulaCSModel)

Fit an GLMCopulaCSModel object by MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.ρ, gcm.σ2.

fit!(gcm::GaussianCopulaARModel)

Fit an GaussianCopulaARModel object by MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.τ, gcm.ρ, gcm.σ2.

fit!(gcm::GLMCopulaVCModel, solver=Ipopt.IpoptSolver(print_level=5))

Fit an GLMCopulaVCModel object by MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.Σ, this is for Poisson and Bernoulli base with no additional parameters than the mean.

fit!(gcm::NBCopulaVCModel, solver=Ipopt.IpoptSolver(print_level=5))

Fit an NBCopulaVCModel object by block MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.Σ, gcm.r. In our block updates, we fit 10 iterations of gcm.β, gcm.Σ using IPOPT, followed by 10 iterations of Newton on nuisance parameter gcm.r. Convergence is declared when difference of successive loglikelihood is less than tol.

fit_quasi!(gcm::GaussianCopulaVCModel, solver=Ipopt.IpoptSolver(print_level=5))

Fit an GaussianCopulaVCModel object by MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.Σ, gcm.τ this is for Normal base.

Missing docstring.

Missing docstring for init_β!. Check Documenter's build log for details.

Models

GLMCopula.GLMCopulaVCModel — Type

GLMCopulaVCModel GLMCopulaVCModel(gcs) Gaussian copula variance component model, which contains a vector of GLMCopulaVCObs as data, model parameters, and working arrays.