API

Documentation for QuasiCopua.jl's functions.

Index

QuasiCopula.AR_model

QuasiCopula.CS_model

QuasiCopula.VC_model

QuasiCopula.fit!

QuasiCopula.get_CI

QuasiCopula.logl

Functions

QuasiCopula.fit! — Function

fit!(gcm::GLMCopulaARModel, solver=Ipopt.IpoptSolver)

Fit an GLMCopulaARModel object by MLE using a nonlinear programming solver. This is for Poisson and Bernoulli base distributions with no additional base distribution parameters than the mean. Start point should be provided in gcm.β, gcm.ρ, gcm.σ2.

Arguments

gcm: A GLMCopulaARModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level=3, max_iter = 100, tol = 10^-6, limited_memory_max_history = 20, hessian_approximation = "limited-memory"))

fit!(gcm::GLMCopulaCSModel, solver=Ipopt.IpoptSolver)

Fit an GLMCopulaCSModel object by MLE using a nonlinear programming solver. This is for Poisson and Bernoulli base distributions with no additional base distribution parameters than the mean. Start point should be provided in gcm.β, gcm.ρ, gcm.σ2.

Arguments

gcm: A GLMCopulaCSModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level=3, max_iter = 100, tol = 10^-6, limited_memory_max_history = 20, hessian_approximation = "limited-memory"))

fit!(gcm::Union{GaussianCopulaARModel, GaussianCopulaCSModel}, solver=Ipopt.IpoptSolver)

Fit a GaussianCopulaARModel or GaussianCopulaCSModel object by MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.τ, gcm.ρ, gcm.σ2.

Arguments

gcm: A GaussianCopulaARModel or GaussianCopulaCSModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level=3, max_iter = 100, tol = 10^-6, limited_memory_max_history = 20, hessian_approximation = "limited-memory"))

fit!(gcm::Union{GLMCopulaVCModel, Poisson_Bernoulli_VCModel}, solver=Ipopt.IpoptSolver(print_level=5))

Fit a GLMCopulaVCModel or Poisson_Bernoulli_VCModel model object by MLE using a nonlinear programming solver. This is for Poisson and Bernoulli base distributions or a mixture of the two with no additional base distribution parameters than the mean. Start point should be provided in gcm.β, gcm.θ.

Arguments

gcm: A GLMCopulaVCModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level=3, max_iter = 100, tol = 10^-6, limited_memory_max_history = 20, hessian_approximation = "limited-memory"))

fit!(gcm::NBCopulaVCModel, solver=Ipopt.IpoptSolver)

Fit an NBCopulaVCModel object by block MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.θ, gcm.r. In our block updates, we fit 15 iterations of gcm.β, gcm.θ using IPOPT, followed by 10 iterations of Newton on nuisance parameter gcm.r. Convergence is declared when difference of successive loglikelihood is less than tol.

Arguments

gcm: A NBCopulaVCModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level = 0, max_iter = 15, limited_memory_max_history = 20, warm_start_init_point = "yes", mu_strategy = "adaptive", hessian_approximation = "limited-memory"))

Optional Arguments

tol: Convergence tolerance for the max block iter updates (default tol = 1e-6).

maxBlockIter: Number of maximum block iterations to update gcm.β, gcm.θ and gcm.r (default maxBlockIter = 10).

fit!(gcm::NBCopulaARModel, solver=Ipopt.IpoptSolver)

Fit an NBCopulaARModel object by block MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.ρ, gcm.σ2, gcm.r. In our block updates, we fit 15 iterations of gcm.β, gcm.ρ, gcm.σ2 using IPOPT, followed by 10 iterations of Newton on nuisance parameter gcm.r. Convergence is declared when difference of successive loglikelihood is less than tol.

Arguments

gcm: A NBCopulaARModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level = 0, max_iter = 15, limited_memory_max_history = 20, warm_start_init_point = "yes", mu_strategy = "adaptive", hessian_approximation = "limited-memory"))

Optional Arguments

tol: Convergence tolerance for the max block iter updates (default tol = 1e-6).

maxBlockIter: Number of maximum block iterations to update gcm.β, gcm.θ and gcm.r (default maxBlockIter = 10).

fit!(gcm::NBCopulaCSModel, solver=Ipopt.IpoptSolver)

Fit an NBCopulaCSModel object by block MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.ρ, gcm.σ2, gcm.r. In our block updates, we fit 15 iterations of gcm.β, gcm.ρ, gcm.σ2 using IPOPT, followed by 10 iterations of Newton on nuisance parameter gcm.r. Convergence is declared when difference of successive loglikelihood is less than tol.

Arguments

gcm: A NBCopulaCSModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level = 0, max_iter = 15, limited_memory_max_history = 20, warm_start_init_point = "yes", mu_strategy = "adaptive", hessian_approximation = "limited-memory"))

Optional Arguments

tol: Convergence tolerance for the max block iter updates (default tol = 1e-6).

maxBlockIter: Number of maximum block iterations to update gcm.β, gcm.θ and gcm.r (default maxBlockIter = 10).

fit_quasi!(gcm::GaussianCopulaVCModel, solver=Ipopt.IpoptSolver)

Fit an GaussianCopulaVCModel object by MLE using a nonlinear programming solver. Start point should be provided in gcm.β, gcm.θ, gcm.τ this is for Normal base.

Arguments

gcm: A GaussianCopulaVCModel model object.

solver: Specified solver to use. By default we use IPOPT with 100 quas-newton iterations with convergence tolerance 10^-6. (default solver = Ipopt.IpoptSolver(print_level=3, max_iter = 100, tol = 10^-6, limited_memory_max_history = 20, warm_start_init_point="yes", hessian_approximation = "limited-memory"))

QuasiCopula.logl — Function

logl(gcm)

Get the loglikelihood at the given parameters in gcm, at the optimal solution.

Arguments

gcm: One of GaussianCopulaVCModel, GaussianCopulaARModel, GaussianCopulaCSModel, GLMCopulaVCModel, GLMCopulaARModel, GLMCopulaCSModel, NBCopulaVCModel, NBCopulaARModel, NBCopulaCSModel or Poisson_Bernoulli_VCModel model objects.

QuasiCopula.get_CI — Function

get_CI(gcm)

Get the confidence interval of all parameters, at the optimal solution.

Arguments

gcm: One of GaussianCopulaVCModel, GaussianCopulaARModel, GaussianCopulaCSModel, GLMCopulaVCModel, GLMCopulaARModel, GLMCopulaCSModel, NBCopulaVCModel, NBCopulaARModel, NBCopulaCSModel or Poisson_Bernoulli_VCModel model objects.

Models

QuasiCopula.AR_model — Function

AR_model(df, y, grouping, d, link; penalized = penalized)

Form the autoregressive (AR(1)) model for intercept only regression with the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance parameter for the AR(1) structured covariance. One can put true (e.g. penalized = true) to add this penalty for numerical stability (default penalized = false).

AR_model(df, y, grouping, covariates, d, link; penalized = penalized)

Form the autoregressive (AR(1)) model for regression with the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

covariates: Covariate names of interest as a vector of Symbols. Each variable name must be present in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance parameter for the AR(1) structured covariance. One can put true (e.g. penalized = true) to add this penalty for numerical stability (default penalized = false).

QuasiCopula.CS_model — Function

CS_model(df, y, grouping, d, link)

Form the compound symmetric (CS) model for intercept only regression with the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance parameter for the CS structured covariance. One can turn this option on by specifying penalized = true to add this penalty for numerical stability. (default penalized = false).

CS_model(df, y, grouping, covariates, d, link; penalized = penalized)

Form the compound symmetric (CS) model for regression with the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

covariates: Covariate names of interest as a vector of Symbols. Each variable name must be present in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance parameter for the CS structured covariance. One can turn this option on by specifying penalized = true to add this penalty for numerical stability. (default penalized = false).

QuasiCopula.VC_model — Function

VC_model(df, y, grouping, d, link)

Form the variance component model (VCM) for intercept only regression with a random intercept covariance matrix and the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance components vector. One can put true (e.g. penalized = true) to add this penalty for numerical stability (default penalized = false).

VC_model(df, y, grouping, covariates, d, link)

Form the variance component model (VCM) for regression with a random intercept covariance matrix and the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

covariates: Covariate names of interest as a vector of Symbols. Each variable name must be present in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance components vector. One can put true (e.g. penalized = true) to add this penalty for numerical stability (default penalized = false).

VC_model(df, y, grouping, covariates, V, d, link)

Form the variance component model (VCM) for intercept only regression with the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

V: Vector of Vector of Positive Semi-Definite (PSD) Covariance Matrices. V is of length n, where n is the number of groups/clusters. Each element of V is also a Vector, but of length m. Here m is the number of variance components. Each element of V is a Vector of di x di PSD covariance matrices under the VCM framework, where d_i is the cluster size of the ith cluster, which may vary for each cluster of observations i in [1, n]. Each of these dimensions must match that specified in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance components vector. One can put true (e.g. penalized = true) to add this penalty for numerical stability (default penalized = false).

VC_model(df, y, grouping, covariates, V, d, link)

Form the variance component model (VCM) for regression with the specified base distribution (d) and link function (link).

Arguments

df: A named DataFrame

y: Ouctcome variable name of interest, specified as a Symbol. This variable name must be present in df.

grouping: Grouping or Clustering variable name of interest, specified as a Symbol. This variable name must be present in df.

covariates: Covariate names of interest as a vector of Symbols. Each variable name must be present in df.

V: Vector of Vector of Positive Semi-Definite (PSD) Covariance Matrices. V is of length n, where n is the number of groups/clusters. Each element of V is also a Vector, but of length m. Here m is the number of variance components. Each element of V is a Vector of di x di PSD covariance matrices under the VCM framework, where d_i is the cluster size of the ith cluster, which may vary for each cluster of observations i in [1, n]. Each of these dimensions must match that specified in df.

d: Base Distribution of outcome from Distributions.jl.

link: Canonical Link function of the base distribution specified in d, from GLM.jl.

Optional Arguments

penalized: Boolean to specify whether or not to add an L2 Ridge penalty on the variance components vector. One can put true (e.g. penalized = true) to add this penalty for numerical stability (default penalized = false).